Move dateutil to timeutil
Also add test.
