+++ mathutil/mathutil.go
